Add ShapeComponent and move some ComponentPool init code to constructor
I moved the resizing of the sparse and mirror arrays from the init to
constructor for the ComponentPool because sometimes I forget to call the
init function and that leads to subscript out of range errors. Now I
won't get that error if I forget, and at worst I get some performance
hiccups as the game warms up but I can fix that by being more diligent
about calling the init function.

+++ GameTest/App/main.cpp
@@ -23,7 +23,7 @@ int WINDOW_HEIGHT = APP_INIT_WINDOW_HEIGHT;
 HWND MAIN_WINDOW_HANDLE = nullptr;
 
 //---------------------------------------------------------------------------------
-static const double FIXED_TIME_STEP = ((1.0 / APP_MAX_FRAME_RATE)*1000.0);
+static const float FIXED_TIME_STEP = ( (1.0f / APP_MAX_FRAME_RATE) * 1000.0f );
 //---------------------------------------------------------------------------------
 // Internal globals for timing.
 double gPCFreq = 0.0;
@@ -132,7 +132,7 @@ void Idle()
 	if (accumulatedTime >= FIXED_TIME_STEP)
 	{
 		gUpdateDeltaTime.Stop();
-		accumulatedTime -= FIXED_TIME_STEP;
+		accumulatedTime -= static_cast<double>(FIXED_TIME_STEP);
 		glutPostRedisplay(); //every time you are done
 		CSimpleControllers::GetInstance().Update();
